# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NPR Seminar NestJS API
  version: 1.0.0
  description: |
    Authoritative contract for the single NestJS backend used by the NPR Seminar
    web and scanner clients.

    ADMIN and SCANNER actors use a Redis-backed project session cookie. Every
    state-changing session request requires X-CSRF-Token, an allowed same-origin
    Origin, and application/json. Public family access uses a separate,
    contact-bound X-Booking-Proof credential. QR tokens, booking proofs, OTPs,
    cookies, passwords, pairing codes, and full parent contact values are secrets and
    MUST be redacted from application, proxy, tracing, analytics, and audit logs.

    Durable resources are never hard-deleted. Booking, check-in, student, sync,
    circuit, and scanner-device histories are append-only and returned in stable
    chronological order.
  x-contract-authority: packages/contracts/openapi.yaml
  x-backend-runtime: NestJS modular monolith
  x-sensitive-log-redaction:
    - Cookie
    - Set-Cookie
    - X-CSRF-Token
    - X-Booking-Proof
    - bookingProof
    - X-QR-Token
    - qrToken
    - pairingCode
    - oneTimeCode
    - contact
    - motherPhone
    - fatherPhone
    - password
  x-canonical-student-snapshot:
    snapshotDate: 2026-07-17
    classNameAdmissionPolicy:
      normalizeForDecision: NFKC_TRIM
      preserveAcceptedSourceValue: true
      allowBracketless: true
      allowOneTerminalScheduleSuffix: true
...
      rejectAsterisk: true
...
    scope: ALL
...
servers:
...
paths:
...
  /api/v1/public/seminar-sessions:
    get:
      operationId: listPublicSeminarSessions
      tags: [Public seminar sessions]
      summary: List sessions currently available to the public booking flow
      description: |
        Returns only published seminars and sessions in their public booking
        window. ALL sessions have branch=null and accept a family from any one
        proof-selected campus. BRANCH sessions have a concrete branch and the
        proof-selected campus must match it. Only sessions whose parent seminar
        is PUBLISHED are bookable even if a session UUID is guessed directly.
      security: []
      parameters:
        - $ref: '#/components/parameters/BranchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Booking-eligible sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSeminarSessionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/admin/seminars:
    get:
      operationId: listAdminSeminars
      tags: [Admin seminars]
      summary: List seminars including archived records
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SeminarStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Seminar page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeminarPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createAdminSeminar
      tags: [Admin seminars]
      summary: Create a seminar
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeminarCreateRequest'
      responses:
        '201':
          description: Seminar created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seminar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/seminars/{seminarId}:
    parameters:
      - $ref: '#/components/parameters/SeminarId'
    get:
      operationId: getAdminSeminar
      tags: [Admin seminars]
      summary: Get a seminar
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Seminar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seminar'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAdminSeminar
      tags: [Admin seminars]
      summary: Update seminar metadata or publication state
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeminarUpdateRequest'
      responses:
        '200':
          description: Updated seminar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seminar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: archiveAdminSeminar
      tags: [Admin seminars]
      summary: Soft-archive a seminar
      description: Transitions status to ARCHIVED; the seminar and all history remain durable.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-hard-delete: forbidden
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveRequest'
      responses:
        '200':
          description: Archived seminar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seminar'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/seminars/{seminarId}/sessions:
    parameters:
      - $ref: '#/components/parameters/SeminarId'
    get:
      operationId: listAdminSeminarSessions
      tags: [Admin seminars]
      summary: List every session and its capacity ledger for a seminar
      description: |
        Every item includes operationsSummary, counted in family-booking
        aggregates rather than seats or participant rows. All session summaries
        are loaded by one status aggregation for the page, never one query per
        session. active is RESERVED plus CHECKED_IN; unchecked is RESERVED.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-operations-summary-query-shape: BATCHED_BY_SESSION_AND_STATUS
      x-operations-summary-count-unit: FAMILY_BOOKING
      x-family-booking-status-semantics:
        active: [RESERVED, CHECKED_IN]
        checkedIn: [CHECKED_IN]
        unchecked: [RESERVED]
        cancelled: [CANCELLED]
        noShow: [NO_SHOW]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Session page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSeminarSessionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createAdminSeminarSession
      tags: [Admin seminars]
      summary: Create a session with an initial capacity
      description: |
        scope and branch are coupled: ALL requires branch=null; BRANCH requires
        one concrete Branch. The current seeded session is ALL with capacity 800.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeminarSessionCreateRequest'
      responses:
        '201':
          description: Session and capacity ledger created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSeminarSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/seminar-sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      operationId: getAdminSeminarSession
      tags: [Admin seminars]
      summary: Get one session and capacity ledger
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Session and capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSeminarSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAdminSeminarSession
      tags: [Admin seminars]
      summary: Update session schedule, state, booking window, or capacity
      description: |
        Capacity cannot be lowered below reservedCount. The update uses the
        supplied version as optimistic concurrency control and never rewrites
        historical booking or check-in snapshots. When scope or branch changes,
        both fields must be supplied: ALL requires branch=null and BRANCH requires
        one concrete Branch. Existing bookings and scanner locks that would
        violate the resulting scope make the transition conflict.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeminarSessionUpdateRequest'
      responses:
        '200':
          description: Updated session and capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSeminarSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: archiveAdminSeminarSession
      tags: [Admin seminars]
      summary: Soft-archive an unused session
      description: |
        Transitions the session to ARCHIVED. Existing or historical reservations
        prevent archival where chronology would be obscured. No row is deleted.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-hard-delete: forbidden
      x-mutation-kind: durable
      x-idempotency: required
      x-requires-csrf: true
      x-requires-same-origin: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveRequest'
      responses:
        '200':
          description: Archived session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSeminarSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: npr_seminar_session
      description: |
        Opaque project-specific Redis session cookie for ADMIN and SCANNER.
        Production attributes are HttpOnly, Secure, SameSite=Lax, Path=/, with
        no Domain where possible. Cookie values are never logged.
    bookingProof:
      type: apiKey
      in: header
      name: X-Booking-Proof
      description: |
        Short-lived 256-bit base64url bearer credential bound to a normalized
        family contact digest, purpose-derived scopes, and expiry. PostgreSQL
        stores SHA-256(proof) only. The raw value is returned once, is independent
        from the project session cookie, and must be redacted from every log.
    qrCredential:
      type: apiKey
      in: header
      name: X-QR-Token
      description: |
        Opaque family QR credential used only for the minimal public pass.
        The raw value is secret, never persisted, and redacted from all logs.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated UUID scoped to actor and operation. Same key and same
        canonical request replays the stored secret-free result; same key with a
        different request returns 409. Raw one-time QR or pairing secrets are
        never present in replay snapshots.
      schema:
        type: string
        format: uuid
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Session-bound CSRF token. The value is secret and MUST NOT be logged.
      schema:
        type: string
        minLength: 32
        maxLength: 512
    Origin:
      name: Origin
      in: header
      required: true
      description: Exact configured browser origin; server also validates Host and trusted proxy metadata.
      schema:
        type: string
        format: uri
        maxLength: 300
...
    SeminarId:
      name: seminarId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BranchQuery:
      name: branch
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Branch'
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: Malformed input, invalid header, or invalid confirmation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, invalid, or revoked session/proof/credential
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Role, CSRF, same-origin, contact ownership, or scope check failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found or intentionally concealed from this actor
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State transition, capacity, duplicate, lock, version, or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: Required PostgreSQL, Redis, source, or cryptographic configuration is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Branch:
      type: string
      enum: [CAMPUS_A, CAMPUS_B, CAMPUS_C]
      x-labels:
        CAMPUS_A: A
        CAMPUS_B: B
        CAMPUS_C: C
...
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
...
    SeminarStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    Seminar:
      type: object
      required: [seminarId, title, description, status, version, createdAt, updatedAt]
      properties:
        seminarId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
        status:
          $ref: '#/components/schemas/SeminarStatus'
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    SeminarCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
      additionalProperties: false
    SeminarUpdateRequest:
      type: object
      required: [expectedVersion]
      minProperties: 2
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 5000
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
        expectedVersion:
          type: integer
          minimum: 1
      additionalProperties: false
    ArchiveRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 3
          maxLength: 500
      additionalProperties: false
    SeminarPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Seminar'
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
    SeminarSessionScope:
      type: string
      enum: [ALL, BRANCH]
      description: |
        ALL is shared by CAMPUS_A, CAMPUS_B, and CAMPUS_C and requires branch=null.
        BRANCH is restricted to exactly one concrete branch.
    SeminarSessionStatus:
      type: string
      enum: [DRAFT, OPEN, CLOSED, CANCELLED, ARCHIVED]
    PublicAvailability:
      type: string
      enum: [AVAILABLE, FULL, NOT_OPEN, CLOSED]
    PublicSeminarSession:
      type: object
      required:
        - seminarId
        - seminarSessionId
        - seminarTitle
        - scope
        - branch
        - startsAt
        - endsAt
        - location
        - bookingOpensAt
        - bookingClosesAt
        - guestBookingEnabled
        - availability
        - remainingCapacity
      properties:
        seminarId:
          type: string
          format: uuid
        seminarSessionId:
          type: string
          format: uuid
        seminarTitle:
          type: string
        scope:
          $ref: '#/components/schemas/SeminarSessionScope'
        branch:
          description: Null exactly when scope is ALL; concrete exactly when scope is BRANCH.
          oneOf:
            - $ref: '#/components/schemas/Branch'
            - type: 'null'
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          maxLength: 300
        bookingOpensAt:
          type: string
          format: date-time
        bookingClosesAt:
          type: string
          format: date-time
        guestBookingEnabled:
          type: boolean
          description: Whether a contact with no active enrolled student in any campus may use the guest flow.
        availability:
          $ref: '#/components/schemas/PublicAvailability'
        remainingCapacity:
          type: integer
          minimum: 0
      allOf:
        - if:
            properties:
              scope:
                const: ALL
            required: [scope]
          then:
            properties:
              branch:
                type: 'null'
        - if:
            properties:
              scope:
                const: BRANCH
            required: [scope]
          then:
            properties:
              branch:
                $ref: '#/components/schemas/Branch'
      additionalProperties: false
    PublicSeminarSessionPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicSeminarSession'
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
...
    AdminSeminarSession:
      type: object
      required:
        - seminarSessionId
        - seminarId
        - scope
        - branch
        - startsAt
        - endsAt
        - location
        - bookingOpensAt
        - bookingClosesAt
        - status
        - guestBookingEnabled
        - capacity
        - version
        - createdAt
        - updatedAt
      properties:
        seminarSessionId:
          type: string
          format: uuid
        seminarId:
          type: string
          format: uuid
        scope:
          $ref: '#/components/schemas/SeminarSessionScope'
        branch:
          description: Null exactly when scope is ALL; concrete exactly when scope is BRANCH.
          oneOf:
            - $ref: '#/components/schemas/Branch'
            - type: 'null'
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          minLength: 1
          maxLength: 300
        bookingOpensAt:
          type: string
          format: date-time
        bookingClosesAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SeminarSessionStatus'
        guestBookingEnabled:
          type: boolean
        capacity:
          $ref: '#/components/schemas/SessionCapacity'
        operationsSummary:
          $ref: '#/components/schemas/SessionOperationsSummary'
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      allOf:
        - if:
            properties:
              scope:
                const: ALL
            required: [scope]
          then:
            properties:
              branch:
                type: 'null'
        - if:
            properties:
              scope:
                const: BRANCH
            required: [scope]
          then:
            properties:
              branch:
                $ref: '#/components/schemas/Branch'
      additionalProperties: false
    AdminSeminarSessionListItem:
      description: Admin session projection with the required batched family-booking operations summary used by the list endpoint.
      allOf:
        - $ref: '#/components/schemas/AdminSeminarSession'
        - type: object
          required: [operationsSummary]
    AdminSeminarSessionPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminSeminarSessionListItem'
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
    SeminarSessionCreateRequest:
      type: object
      required:
        - scope
        - branch
        - startsAt
        - endsAt
        - location
        - bookingOpensAt
        - bookingClosesAt
        - capacity
      properties:
        scope:
          $ref: '#/components/schemas/SeminarSessionScope'
        branch:
          description: Must be null for ALL and concrete for BRANCH.
          oneOf:
            - $ref: '#/components/schemas/Branch'
            - type: 'null'
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          minLength: 1
          maxLength: 300
        bookingOpensAt:
          type: string
          format: date-time
        bookingClosesAt:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 1
          maximum: 100000
        guestBookingEnabled:
          type: boolean
          default: false
      allOf:
        - if:
            properties:
              scope:
                const: ALL
            required: [scope]
          then:
            properties:
              branch:
                type: 'null'
        - if:
            properties:
              scope:
                const: BRANCH
            required: [scope]
          then:
            properties:
              branch:
                $ref: '#/components/schemas/Branch'
      additionalProperties: false
    SeminarSessionUpdateRequest:
      type: object
      required: [expectedVersion]
      minProperties: 2
      properties:
        scope:
          $ref: '#/components/schemas/SeminarSessionScope'
        branch:
          description: Must be supplied with scope; null for ALL and concrete for BRANCH.
          oneOf:
            - $ref: '#/components/schemas/Branch'
            - type: 'null'
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          minLength: 1
          maxLength: 300
        bookingOpensAt:
          type: string
          format: date-time
        bookingClosesAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [DRAFT, OPEN, CLOSED, CANCELLED]
        capacity:
          type: integer
          minimum: 1
          maximum: 100000
        guestBookingEnabled:
          type: boolean
        expectedVersion:
          type: integer
          minimum: 1
      dependentRequired:
        scope: [branch]
        branch: [scope]
      allOf:
        - if:
            properties:
              scope:
                const: ALL
            required: [scope]
          then:
            properties:
              branch:
                type: 'null'
        - if:
            properties:
              scope:
                const: BRANCH
            required: [scope]
          then:
            properties:
              branch:
                $ref: '#/components/schemas/Branch'
      additionalProperties: false
...